KCenters.jl

API

KCenters.AbstractCenterSelection

KCenters.CentroidSelection

KCenters.ClusteringData

KCenters.KnnCentroidSelection

KCenters.associate_centroids

KCenters.change_criterion

KCenters.dnet

KCenters.dnet

KCenters.enet

KCenters.epsilon_criterion

KCenters.fftraversal

KCenters.fun_criterion

KCenters.invindex

KCenters.kcenters

KCenters.kcenters

KCenters.knr

KCenters.partition

KCenters.salesman_criterion

KCenters.sequence

KCenters.size_criterion

KCenters.softmax!

KCenters.transform

KCenters.AbstractCenterSelection — Type

abstract type AbstractCenterSelection end

Abstract type for all center selection strategies

KCenters.CentroidSelection — Type

CentroidSelection()
center(sel::CentroidSelection, lst::AbstractVector)

Computes a center as the centroid of the lst set of points

KCenters.ClusteringData — Type

struct ClusteringData{DataType<:AbstractVector}
    # n elements in the dataset, m centers
    centers::DataType # centers, m entries
    freqs::Vector{Int32} # number of elements associated to each center, m entries
    dmax::Vector{Float32} # stores the distant element associated to each center, m entries
    codes::Vector{Int32} # id of the associated center, n entries
    distances::Vector{Float32} # from each element to its nearest center (label), n entries
    err::Vector{Float32} # dynamic of the error function, at least one entry
end

The datastructure output of our clustering procedures

KCenters.KnnCentroidSelection — Type

KnnCentroidSelection(sel1::AbstractCenterSelection, sel2::AbstractCenterSelection, dist::PreMetric, k::Int32)
KnnCentroidSelection(; sel1=CentroidSelection(), sel2=CentroidSelection(), dist=SqL2Distance(), k=0) = KnnCentroidSelection(sel, dist, convert(Int32, k))
center(sel::KnnCentroidSelection, lst::AbstractVector)

Computes a center using the sel1 selection strategy, and computes the final center over the set of k nearest neighbors of the initial center (from lst) using the dist distance function.

KCenters.associate_centroids — Method

associate_centroids(dist::PreMetric, X, centers)

Returns the named tuple (codes=codes, freqs=freqs, distances=distances, err=s) where codes contains the nearest center index for each item in X under the context of the dist distance function. C is the set of centroids and X the dataset of objects. C also can be provided as a SimilaritySearch's Index.

KCenters.change_criterion — Function

change_criterion(tol=0.001, window=3)

Creates a fuction that stops the process whenever the maximum distance converges (averaging window far items). The tol parameter defines the tolerance range.

KCenters.dnet — Method

dnet(dist::PreMetric, X::AbstractVector{T}, numcenters::Integer) where T

Selects numcenters far from each other based on density nets.

dist distance function

X the objects to be computed

numcenters number of centers to be computed

Returns a named tuple $(nn, irefs, dmax)$.

irefs contains the list of centers (indexes to $X$)

seq contains the $k$ nearest references for each object in $X$ (in $X$ order)

dmax a list of coverage-radius of each center (aligned with irefs centers) smallest distance among centers

KCenters.dnet — Method

dnet(callback::Function, dist::PreMetric, X::AbstractVector{T}, k::Integer) where {T}

A k-net is a set of points M such that each object in X can be:

It is in M

It is in the knn set of an object in M (defined with the distance function dist)

The size of M is determined by $\leftceil|X|/k\rightceil$

The dnet function uses the callback function as an output mechanism. This function is called on each center as callback(centerId, res) where res is a KnnResult object (from SimilaritySearch.jl).

KCenters.enet — Method

enet(dist::PreMetric, X::AbstractVector{T}, numcenters::Int, knr::Int=1; verbose=false) where T

Selects numcenters far from each other based on Farthest First Traversal.

dist distance function

X the objects to be computed

numcenters number of centers to be computed

knr number of nearest references per object (knr=1 defines a partition)

Returns a named tuple $(nn, irefs, dmax)$.

irefs contains the list of centers (indexes to $X$)

seq contains the $k$ nearest references for each object in $X$ (in $X$ order)

dmax smallest distance among centers

KCenters.epsilon_criterion — Method

epsilon_criterion(e)

Creates a function that evaluates the stop criterion when the distance between far items achieves the given e

KCenters.fftraversal — Method

fftraversal(callback::Function, dist::PreMetric, X::AbstractVector{T}, stop, callbackdist=_ignore3) where {T}

Selects a number of farthest points in X, using a farthest first traversal

The callback function is called on each selected far point as callback(centerID, dmax) where dmax is the distance to the nearest previous reported center (the first is reported with typemax)

The selected objects are far under the dist distance function with signature (T, T) -> Float64

The number of points is determined by the stop criterion function with signature (Float64[], T[]) -> Bool

The first argument corresponds to the list of known distances (far objects)

The second argument corresponds to the database

Check criterions.jl for basic implementations of stop criterions

The callbackdist function is called on each distance evaluation between pivots and items in the dataset callbackdist(index-pivot, index-item, distance)

KCenters.fun_criterion — Method

fun_criterion(fun::Function)

Creates a stop-criterion function that stops whenever the number of far items reaches $\lceil fun(|database|)\rceil$. Already defined examples:

    sqrt_criterion() = fun_criterion(sqrt)
    log2_criterion() = fun_criterion(log2)

KCenters.invindex — Method

invindex(objects::AbstractVector{T}, refs::AbstractSearchContext; k::Int=1) where T

Creates an inverted index from references to objects. So, an object $u$ is in $r$'s posting list iff $r$ is among the $k$ nearest references of $u$.

KCenters.kcenters — Method

kcenters(dist::PreMetric, X::AbstractVector{T}, y::CategoricalArray, sel::AbstractCenterSelection=CentroidSelection()) where T

Computes a center per region (each region is defined by the set of items having the same label in y). The output is compatible with kcenters function when eltype(y) is Int

KCenters.kcenters — Method

kcenters(dist::PreMetric, X::AbstractVector{T}, k::Integer; sel::AbstractCenterSelection=CentroidSelection(), initial=:fft, maxiters=0, tol=0.001, recall=1.0) where T
kcenters(dist::PreMetric, X::AbstractVector{T}, C::AbstractzVector{T}; sel::AbstractCenterSelection=CentroidSelection(), maxiters=30, tol=0.001, recall=1.0) where T

Performs a kcenters clustering of X using dist as distance function and sel to compute center objects. It is based on the Lloyd's algorithm yet using different algorithms as initial clusters. - :fft the farthest first traversal selects a set of farthest points among them to serve as cluster seeds. - :dnet the density net algorithm selects a set of points following the same distribution of the datasets; in contrast with a random selection, :dnet ensures that the selected points are not $\lfloor n/k \rfloor$ nearest neighbors. - :sfft the :fft over a $k + \log n$ random sample - :sdnet the :dnet over a $k + \log n$ random sample - :rand selects the set of random points along the dataset.

If recall is 1.0 then an exhaustive search is made to find associations of each item to its nearest cluster; if $0 < recall < 0$ then an approximate index (SearchGraph from SimilaritySearch.jl) will be used for the same purpose; the recall controls the expected search quality (trade with search time).

KCenters.knr — Method

knr(objects::AbstractVector{T}, refs::AbstractSearchContext) where T

Computes an array of k-nearest neighbors for objects

KCenters.partition — Method

partition(callback::Function, objects::AbstractVector{T}, refs::AbstractSearchContext; k::Int=1) where T

Groups items in objects using a nearest neighbor rule over refs. The output is controlled using a callback function. The call is performed in objects order.

callback is a function that is called for each (objID, refItem)

objects is the input dataset

dist a distance function $(T, T) \rightarrow \mathbb{R}$

refs the list of references

k specifies the number of nearest neighbors to use

indexclass specifies the kind of index to be used, a function receiving (refs, dist) as arguments, and returning the new metric index

Please note that each object can be related to more than one group $k > 1$ (default $k=1$).

KCenters.salesman_criterion — Method

salesman_criterion()

It creates a function that explores the entire dataset making a full farthest first traversal approximation

KCenters.sequence — Method

sequence(objects::AbstractVector{T}, refs::AbstractSearchContext) where T

Computes the nearest reference of each item in the dataset and return it as a sequence of identifiers

KCenters.size_criterion — Method

size_criterion(maxsize)

Creates a function that stops when the number of far items are equal or larger than the given maxsize

KCenters.softmax! — Method

softmax!(vec::AbstractVector)

Inline computation of the softmax function on the input vector

KCenters.transform — Method

transform(centers::AbstractVector{T}, dmax::AbstractVector, kernel::Function, q::T) where T
transform(centers::AbstractVector{T}, dmax::AbstractVector, kernel::Function, queries::AbstractVector{T}, normalize!::Function=identity) where T

Creates a vector representation of q being its projection along the centrois in vor using the kernel function to perform the projection at each center. If a collection of Ts is given, then it computes transform for each one.