Center selection schemes

KCenters support different ways to compute or select object prototypes (centers). You can use them through center function; any new scheme must specialize from AbstractCenterSelection and specialize center function.

KCenters.AbstractCenterSelection — Type

abstract type AbstractCenterSelection end

Abstract type for all center selection strategies

KCenters.CentroidSelection — Type

CentroidSelection()
center(sel::CentroidSelection, lst::AbstractVector)

Computes a center as the centroid of the lst set of points

Missing docstring.

Missing docstring for RandomCenterSelection. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MedoidSelection. Check Documenter's build log for details.

KCenters.KnnCentroidSelection — Type

KnnCentroidSelection(sel1::AbstractCenterSelection, sel2::AbstractCenterSelection, dist::SemiMetric, k::Int32)
KnnCentroidSelection(; sel1=CentroidSelection(), sel2=CentroidSelection(), dist=SqL2Distance(), k=0) = KnnCentroidSelection(sel, dist, convert(Int32, k))
center(sel::KnnCentroidSelection, lst::AbstractVector)

Computes a center using the sel1 selection strategy, and computes the final center over the set of k nearest neighbors of the initial center (from lst) using the dist distance function.

Missing docstring.

Missing docstring for center. Check Documenter's build log for details.