Clustering

Functions to cluster objects using distance functions and center selection schemes.

KCenters.ClusteringData — Type

struct ClusteringData{DataType<:AbstractVector}
    # n elements in the dataset, m centers
    centers::DataType # centers, m entries
    freqs::Vector{Int32} # number of elements associated to each center, m entries
    dmax::Vector{Float32} # stores the distant element associated to each center, m entries
    codes::Vector{Int32} # id of the associated center, n entries
    distances::Vector{Float32} # from each element to its nearest center (label), n entries
    err::Vector{Float32} # dynamic of the error function, at least one entry
end

The datastructure output of our clustering procedures

KCenters.kcenters — Function

kcenters(dist::PreMetric, X::AbstractVector{T}, y::CategoricalArray, sel::AbstractCenterSelection=CentroidSelection()) where T

Computes a center per region (each region is defined by the set of items having the same label in y). The output is compatible with kcenters function when eltype(y) is Int

kcenters(dist::PreMetric, X::AbstractVector{T}, k::Integer; sel::AbstractCenterSelection=CentroidSelection(), initial=:fft, maxiters=0, tol=0.001, recall=1.0) where T
kcenters(dist::PreMetric, X::AbstractVector{T}, C::AbstractzVector{T}; sel::AbstractCenterSelection=CentroidSelection(), maxiters=30, tol=0.001, recall=1.0) where T

Performs a kcenters clustering of X using dist as distance function and sel to compute center objects. It is based on the Lloyd's algorithm yet using different algorithms as initial clusters. - :fft the farthest first traversal selects a set of farthest points among them to serve as cluster seeds. - :dnet the density net algorithm selects a set of points following the same distribution of the datasets; in contrast with a random selection, :dnet ensures that the selected points are not $\lfloor n/k \rfloor$ nearest neighbors. - :sfft the :fft over a $k + \log n$ random sample - :sdnet the :dnet over a $k + \log n$ random sample - :rand selects the set of random points along the dataset.

If recall is 1.0 then an exhaustive search is made to find associations of each item to its nearest cluster; if $0 < recall < 0$ then an approximate index (SearchGraph from SimilaritySearch.jl) will be used for the same purpose; the recall controls the expected search quality (trade with search time).

KCenters.associate_centroids — Function

associate_centroids(dist::PreMetric, X, centers)

Returns the named tuple (codes=codes, freqs=freqs, distances=distances, err=s) where codes contains the nearest center index for each item in X under the context of the dist distance function. C is the set of centroids and X the dataset of objects. C also can be provided as a SimilaritySearch's Index.