Stop criterions for $\epsilon$-net enet

The following functions create an stop criterion function of the form: (dmaxlist, database) -> bool. This function returns true whenever the caller must stop and false otherwise.

KCenters.size_criterion — Function

size_criterion(maxsize)

Creates a function that stops when the number of far items are equal or larger than the given maxsize

KCenters.sqrt_criterion — Function

sqrt_criterion()

Stops after $sqrt(n)$ centers

KCenters.change_criterion — Function

change_criterion(tol=0.001, window=3)

Creates a fuction that stops the process whenever the maximum distance converges (averaging window far items). The tol parameter defines the tolerance range.

KCenters.fun_criterion — Function

fun_criterion(fun::Function)

Creates a stop-criterion function that stops whenever the number of far items reaches $\lceil fun(|database|)\rceil$. Already defined examples:

    sqrt_criterion() = fun_criterion(sqrt)
    log2_criterion() = fun_criterion(log2)

KCenters.log2_criterion — Function

log2_criterion()

Stops after $log_2(n)$ centers

KCenters.epsilon_criterion — Function

epsilon_criterion(e)

Creates a function that evaluates the stop criterion when the distance between far items achieves the given e

KCenters.salesman_criterion — Function

salesman_criterion()

It creates a function that explores the entire dataset making a full farthest first traversal approximation