enet

Functions to compute $\epsilon$-nets

KCenters.enet — Function

enet(dist::SemiMetric, X::AbstractDatabase, stop)

Selects a number of farthest points in X, using a farthest first traversal

enet(dist::SemiMetric, X::AbstractDatabase, k::Integer) = enet(dist, X, size_criterion(k))

Selects k items far from each other based on Farthest First Traversal algorithm.

Returns a named tuple $(nn, irefs, dmax)$.

irefs contains the list of centers (indexes to $X$)

seq contains the nearest references for each object in $X$ (in $X$ order)

dmax smallest distance among centers