Miscelaneous functions

Helping functions for partitioning data.

KCenters.knr — Function

knr(objects::AbstractVector{T}, refs::AbstractSearchContext) where T

Computes an array of k-nearest neighbors for objects

KCenters.sequence — Function

sequence(objects::AbstractVector{T}, refs::AbstractSearchContext) where T

Computes the nearest reference of each item in the dataset and return it as a sequence of identifiers

KCenters.invindex — Function

invindex(objects::AbstractVector{T}, refs::AbstractSearchContext; k::Int=1) where T

Creates an inverted index from references to objects. So, an object $u$ is in $r$'s posting list iff $r$ is among the $k$ nearest references of $u$.

KCenters.partition — Function

partition(callback::Function, objects::AbstractVector{T}, refs::AbstractSearchContext; k::Int=1) where T

Groups items in objects using a nearest neighbor rule over refs. The output is controlled using a callback function. The call is performed in objects order.

callback is a function that is called for each (objID, refItem)

objects is the input dataset

dist a distance function $(T, T) \rightarrow \mathbb{R}$

refs the list of references

k specifies the number of nearest neighbors to use

indexclass specifies the kind of index to be used, a function receiving (refs, dist) as arguments, and returning the new metric index

Please note that each object can be related to more than one group $k > 1$ (default $k=1$).